# Play around with optimizing FastLU