```julia
using LinearSolve, LinearAlgebra, BenchmarkTools

A = rand(4, 4)
b = rand(4)
prob = LinearProblem(A, b)

using BenchmarkTools
@benchmark solve($prob)
@benchmark solve($prob, $(GenericLUFactorization(LinearSolve.RowMaximum())))
@benchmark solve($prob, $(LinearSolve.LUFactorization()))
@benchmark solve($prob, $(FastLUFactorization()))

@profview for i in 1:1000000 solve(prob, LinearSolve.LUFactorization()) end
@profview for i in 1:1000000 solve(prob, FastLUFactorization()) end
```

```julia


BenchmarkTools.Trial: 10000 samples with 187 evaluations.
 Range (min … max):  548.797 ns …  12.246 μs  ┊ GC (min … max): 0.00% … 93.99%
 Time  (median):     567.513 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   605.255 ns ± 444.103 ns  ┊ GC (mean ± σ):  4.52% ±  5.73%

   ▃▅▆████▇▇▅▄▃▃▃▃▃▃▂▂▂▂▁▁▁▂▂▁▁▁ ▁  ▁▁                          ▃
  ▇████████████████████████████████████████████▇▇▇▇▆▆▇▇▄▇▆▆▆▅▄▅ █
  549 ns        Histogram: log(frequency) by time        704 ns <

 Memory estimate: 992 bytes, allocs estimate: 10.

BenchmarkTools.Trial: 10000 samples with 188 evaluations.
 Range (min … max):  537.676 ns …  12.515 μs  ┊ GC (min … max): 0.00% … 94.08%
 Time  (median):     556.516 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   591.203 ns ± 424.057 ns  ┊ GC (mean ± σ):  3.78% ±  4.94%

   ▁▃▅▇██▇▄▂▃▄▃▃▃▃▃▂▂▁▁▁▁ ▁▁▁▁▁▁ ▁                              ▂
  ▅███████████████████████████████████▇▇▇▇▆▇▇▆▇▆▆▆▇▇▅▅▆▄▅▅▆▅▆▅▅ █
  538 ns        Histogram: log(frequency) by time        717 ns <

 Memory estimate: 704 bytes, allocs estimate: 9.

BenchmarkTools.Trial: 10000 samples with 176 evaluations.
 Range (min … max):  601.562 ns …  12.679 μs  ┊ GC (min … max): 0.00% … 93.34%
 Time  (median):     620.977 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   651.407 ns ± 427.306 ns  ┊ GC (mean ± σ):  3.34% ±  4.75%

    ▃▅▇███▇▆▅▃▃▃▃▂▃▁▂▁▁  ▁  ▁                                   ▂
  ▄▇█████████████████████████████▇██████▇▇▇██▇▇▇▆▇▆▆▆▆▅▆▅▅▅▅▅▅▅ █
  602 ns        Histogram: log(frequency) by time        761 ns <

 Memory estimate: 704 bytes, allocs estimate: 9.

BenchmarkTools.Trial: 10000 samples with 150 evaluations.
 Range (min … max):  678.887 ns …  18.727 μs  ┊ GC (min … max): 0.00% … 95.06%
 Time  (median):     697.773 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   738.816 ns ± 635.637 ns  ┊ GC (mean ± σ):  4.30% ±  4.75%

   ▁▃▅▇███▇▆▅▄▃▃▂▂▂▂▁▁▁▁ ▁ ▁  ▁▁▁ ▁ ▁                           ▂
  ▆███████████████████████████████████████▇█▇▇█▇▇▇▇▆▆▆▆▆▅▅▆▅▄▄▅ █
  679 ns        Histogram: log(frequency) by time        834 ns <

 Memory estimate: 848 bytes, allocs estimate: 11.
```