@ChrisRackauckas 2023-06-01
It's not necessarily slow, it just allocates a bigger workspace so for single solves it's slower.