@ChrisRackauckas 2023-06-18
```julia
using LinearSolve, LinearAlgebra, BenchmarkTools

A = rand(4, 4)
b = rand(4)
prob = LinearProblem(A, b)

using BenchmarkTools
@benchmark solve($prob)
@benchmark solve($prob, $(GenericLUFactorization(LinearSolve.RowMaximum())))
@benchmark solve($prob, $(LinearSolve.LUFactorization()))
@benchmark solve($prob, $(FastLUFactorization()))

cache = init(prob;)
@benchmark solve!($cache)
cache = init(prob, GenericLUFactorization(LinearSolve.RowMaximum()));
@benchmark solve!($cache)
cache = init(prob, LinearSolve.LUFactorization());
@benchmark solve!($cache)
cache = init(prob, FastLUFactorization());
@benchmark solve!($cache)


@profview for i in 1:1000000 solve(prob, LinearSolve.LUFactorization()) end
@profview for i in 1:10000000 solve!(cache); end
```

It's at least optimized:

```julia
BenchmarkTools.Trial: 10000 samples with 905 evaluations.
 Range (min … max):  123.425 ns … 387.514 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     129.392 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   130.335 ns ±   5.647 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

              ▇▄█▆▃▅▁▂
  ▂▂▂▂▂▂▂▃▄▅██████████▇▇▅▄▅▄▃▃▃▃▃▃▂▂▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▁▂▂ ▃
  123 ns           Histogram: frequency by time          146 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.


BenchmarkTools.Trial: 10000 samples with 962 evaluations.
 Range (min … max):  86.798 ns … 231.809 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     90.956 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   92.286 ns ±   9.386 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▆█▇▅▃▁                                                      ▂
  ████████▇█▇▅▆▅▅▄▅▄▄▆▄▅▄▄▄▁▃▄▅▃▃▃▃▁▁▃▁▁▁▃▁▃▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▃▄▄ █
  86.8 ns       Histogram: log(frequency) by time       175 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.


BenchmarkTools.Trial: 10000 samples with 962 evaluations.
 Range (min … max):  89.605 ns … 188.462 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     94.075 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   95.076 ns ±   6.761 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

          ▅▇██▆▁
  ▂▂▂▂▃▃▅████████▅▄▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▁▂▂▁▂▂▂▂▂▂▁▁▂▂ ▃
  89.6 ns         Histogram: frequency by time          114 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.


BenchmarkTools.Trial: 10000 samples with 952 evaluations.
 Range (min … max):   96.744 ns … 197.794 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     100.420 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   101.944 ns ±   9.926 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▅█▁
  ▃███▄▃▂▃▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▁▂▂▂▂▂▂▁▂▂▁▁▂▂▁▁▁▁▁▁▁▁▁▂▁▁▁▁▁▁▁▁▁▁▁▂▂▂ ▂
  96.7 ns          Histogram: frequency by time          186 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```